Pushed group-to-zone and zone-to-group state synthesis into the client application controller basis so that all clients, including hlxproxyd, can leverage it.
