+++ src/lib/client/GroupsControllerBasis.cpp
@@ -621,6 +621,37 @@ GroupsControllerBasis :: Query(const Model::GroupModel::IdentifierType &aGroupId
     return (lRetval);
 }
 
+/**
+ *  @brief
+ *    Get the group model associated with specified group
+ *    identifier.
+ *
+ *  @param[in]   aIdentifier  An immutable reference to the group
+ *                            model to obtain.
+ *  @param[out]  aModel       A reference to an immutable pointer
+ *                            by which to return the group model.
+ *
+ *  @retval  kStatus_Success  If successful.
+ *  @retval  -ERANGE          If the group identifier is smaller
+ *                            or larger than supported.
+ *
+ */
+Status
+GroupsControllerBasis :: GetGroup(const GroupModel::IdentifierType &aIdentifier, const GroupModel *&aModel) const
+{
+    Status  lRetval = kStatus_Success;
+
+
+    lRetval = Common::GroupsControllerBasis::ValidateIdentifier(aIdentifier);
+    nlREQUIRE_SUCCESS(lRetval, done);
+
+    lRetval = mGroupsModel.GetGroup(aIdentifier, aModel);
+    nlREQUIRE_SUCCESS(lRetval, done);
+
+ done:
+    return (lRetval);
+}
+
 // MARK: Command Completion Handlers
 
 /**
