+++ src/lib/client/ZonesController.cpp
@@ -264,37 +264,6 @@ done:
 
 /**
  *  @brief
- *    Get the zone model associated with specified zone
- *    identifier.
- *
- *  @param[in]   aIdentifier  An immutable reference to the zone
- *                            model to obtain.
- *  @param[out]  aModel       A reference to an immutable pointer
- *                            by which to return the zone model.
- *
- *  @retval  kStatus_Success  If successful.
- *  @retval  -ERANGE          If the zone identifier is smaller
- *                            or larger than supported.
- *
- */
-Status
-ZonesController :: GetZone(const IdentifierType &aIdentifier, const ZoneModel *&aModel) const
-{
-    Status  lRetval = kStatus_Success;
-
-
-    lRetval = ValidateIdentifier(aIdentifier);
-    nlREQUIRE_SUCCESS(lRetval, done);
-
-    lRetval = mZones.GetZone(aIdentifier, aModel);
-    nlREQUIRE_SUCCESS(lRetval, done);
-
-done:
-    return (lRetval);
-}
-
-/**
- *  @brief
  *    Get the zone identifier with the specified name.
  *
  *  This attempts to lookup the zone identifier for the zone
